Correção de acentuação das imagens e extensão
